Move functions to utils.py
Moving the static functions from routes.py to util.py
